+++ src/whop_sdk/resources/memberships.py
@@ -71,6 +71,7 @@ class MembershipsResource(SyncAPIResource):
         Required permissions:
 
         - `member:basic:read`
+        - `member:email:read`
 
         Args:
           extra_headers: Send extra headers
@@ -109,6 +110,7 @@ class MembershipsResource(SyncAPIResource):
         Required permissions:
 
         - `member:manage`
+        - `member:email:read`
         - `member:basic:read`
 
         Args:
@@ -178,6 +180,7 @@ class MembershipsResource(SyncAPIResource):
         Required permissions:
 
         - `member:basic:read`
+        - `member:email:read`
 
         Args:
           after: Returns the elements in the list that come after the specified cursor.
@@ -269,6 +272,7 @@ class MembershipsResource(SyncAPIResource):
         Required permissions:
 
         - `member:manage`
+        - `member:email:read`
         - `member:basic:read`
 
         Args:
@@ -313,6 +317,7 @@ class MembershipsResource(SyncAPIResource):
         Required permissions:
 
         - `member:manage`
+        - `member:email:read`
         - `member:basic:read`
 
         Args:
@@ -355,6 +360,7 @@ class MembershipsResource(SyncAPIResource):
         Required permissions:
 
         - `member:manage`
+        - `member:email:read`
         - `member:basic:read`
 
         Args:
@@ -370,6 +376,45 @@ class MembershipsResource(SyncAPIResource):
             raise ValueError(f"Expected a non-empty value for `id` but received {id!r}")
         return self._post(
             f"/memberships/{id}/resume",
+            options=make_request_options(
+                extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
+            ),
+            cast_to=Membership,
+        )
+
+    def uncancel(
+        self,
+        id: str,
+        *,
+        # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
+        # The extra values given here take precedence over values defined on the client or passed to this method.
+        extra_headers: Headers | None = None,
+        extra_query: Query | None = None,
+        extra_body: Body | None = None,
+        timeout: float | httpx.Timeout | None | NotGiven = not_given,
+    ) -> Membership:
+        """
+        Uncancels a membership that was scheduled to cancel at period end
+
+        Required permissions:
+
+        - `member:manage`
+        - `member:email:read`
+        - `member:basic:read`
+
+        Args:
+          extra_headers: Send extra headers
+
+          extra_query: Add additional query parameters to the request
+
+          extra_body: Add additional JSON properties to the request
+
+          timeout: Override the client-level default timeout for this request, in seconds
+        """
+        if not id:
+            raise ValueError(f"Expected a non-empty value for `id` but received {id!r}")
+        return self._post(
+            f"/memberships/{id}/uncancel",
             options=make_request_options(
                 extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
             ),
@@ -414,6 +459,7 @@ class AsyncMembershipsResource(AsyncAPIResource):
         Required permissions:
 
         - `member:basic:read`
+        - `member:email:read`
 
         Args:
           extra_headers: Send extra headers
@@ -452,6 +498,7 @@ class AsyncMembershipsResource(AsyncAPIResource):
         Required permissions:
 
         - `member:manage`
+        - `member:email:read`
         - `member:basic:read`
 
         Args:
@@ -521,6 +568,7 @@ class AsyncMembershipsResource(AsyncAPIResource):
         Required permissions:
 
         - `member:basic:read`
+        - `member:email:read`
 
         Args:
           after: Returns the elements in the list that come after the specified cursor.
@@ -612,6 +660,7 @@ class AsyncMembershipsResource(AsyncAPIResource):
         Required permissions:
 
         - `member:manage`
+        - `member:email:read`
         - `member:basic:read`
 
         Args:
@@ -656,6 +705,7 @@ class AsyncMembershipsResource(AsyncAPIResource):
         Required permissions:
 
         - `member:manage`
+        - `member:email:read`
         - `member:basic:read`
 
         Args:
@@ -700,6 +750,7 @@ class AsyncMembershipsResource(AsyncAPIResource):
         Required permissions:
 
         - `member:manage`
+        - `member:email:read`
         - `member:basic:read`
 
         Args:
@@ -715,6 +766,45 @@ class AsyncMembershipsResource(AsyncAPIResource):
             raise ValueError(f"Expected a non-empty value for `id` but received {id!r}")
         return await self._post(
             f"/memberships/{id}/resume",
+            options=make_request_options(
+                extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
+            ),
+            cast_to=Membership,
+        )
+
+    async def uncancel(
+        self,
+        id: str,
+        *,
+        # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
+        # The extra values given here take precedence over values defined on the client or passed to this method.
+        extra_headers: Headers | None = None,
+        extra_query: Query | None = None,
+        extra_body: Body | None = None,
+        timeout: float | httpx.Timeout | None | NotGiven = not_given,
+    ) -> Membership:
+        """
+        Uncancels a membership that was scheduled to cancel at period end
+
+        Required permissions:
+
+        - `member:manage`
+        - `member:email:read`
+        - `member:basic:read`
+
+        Args:
+          extra_headers: Send extra headers
+
+          extra_query: Add additional query parameters to the request
+
+          extra_body: Add additional JSON properties to the request
+
+          timeout: Override the client-level default timeout for this request, in seconds
+        """
+        if not id:
+            raise ValueError(f"Expected a non-empty value for `id` but received {id!r}")
+        return await self._post(
+            f"/memberships/{id}/uncancel",
             options=make_request_options(
                 extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
             ),
@@ -744,6 +834,9 @@ class MembershipsResourceWithRawResponse:
         self.resume = to_raw_response_wrapper(
             memberships.resume,
         )
+        self.uncancel = to_raw_response_wrapper(
+            memberships.uncancel,
+        )
 
 
 class AsyncMembershipsResourceWithRawResponse:
@@ -767,6 +860,9 @@ class AsyncMembershipsResourceWithRawResponse:
         )
         self.resume = async_to_raw_response_wrapper(
             memberships.resume,
+        )
+        self.uncancel = async_to_raw_response_wrapper(
+            memberships.uncancel,
         )
 
 
@@ -792,6 +888,9 @@ class MembershipsResourceWithStreamingResponse:
         self.resume = to_streamed_response_wrapper(
             memberships.resume,
         )
+        self.uncancel = to_streamed_response_wrapper(
+            memberships.uncancel,
+        )
 
 
 class AsyncMembershipsResourceWithStreamingResponse:
@@ -815,4 +914,7 @@ class AsyncMembershipsResourceWithStreamingResponse:
         )
         self.resume = async_to_streamed_response_wrapper(
             memberships.resume,
+        )
+        self.uncancel = async_to_streamed_response_wrapper(
+            memberships.uncancel,
         )