+++ src/whop_sdk/types/app_build_create_params.py
@@ -2,18 +2,13 @@
 
 from __future__ import annotations
 
-from typing import List, Union, Optional
-from typing_extensions import Required, TypeAlias, TypedDict
+from typing import List, Optional
+from typing_extensions import Required, TypedDict
 
 from .shared.app_view_type import AppViewType
 from .shared.app_build_platforms import AppBuildPlatforms
 
-__all__ = [
-    "AppBuildCreateParams",
-    "Attachment",
-    "AttachmentAttachmentInputWithDirectUploadID",
-    "AttachmentAttachmentInputWithID",
-]
+__all__ = ["AppBuildCreateParams", "Attachment"]
 
 
 class AppBuildCreateParams(TypedDict, total=False):
@@ -51,26 +46,11 @@ class AppBuildCreateParams(TypedDict, total=False):
     """
 
 
-class AttachmentAttachmentInputWithDirectUploadID(TypedDict, total=False):
-    """Input for an attachment"""
+class Attachment(TypedDict, total=False):
+    """Attachment input for the app build file.
 
-    direct_upload_id: Required[str]
-    """This ID should be used the first time you upload an attachment.
-
-    It is the ID of the direct upload that was created when uploading the file to S3
-    via the mediaDirectUpload mutation.
+    This should be an upload in .zip format. The zip should contain at least one main_js_bundle.hbc file and optionally an assets folder next to it.
     """
-
-
-class AttachmentAttachmentInputWithID(TypedDict, total=False):
-    """Input for an attachment"""
 
     id: Required[str]
-    """The ID of an existing attachment object.
-
-    Use this when updating a resource and keeping a subset of the attachments. Don't
-    use this unless you know what you're doing.
-    """
-
-
-Attachment: TypeAlias = Union[AttachmentAttachmentInputWithDirectUploadID, AttachmentAttachmentInputWithID]
+    """The ID of an existing file object."""