+++ src/whop_sdk/types/experience_update_params.py
@@ -2,10 +2,10 @@
 
 from __future__ import annotations
 
-from typing import Union, Optional
-from typing_extensions import Literal, Required, TypeAlias, TypedDict
+from typing import Optional
+from typing_extensions import Literal, Required, TypedDict
 
-__all__ = ["ExperienceUpdateParams", "Logo", "LogoAttachmentInputWithDirectUploadID", "LogoAttachmentInputWithID"]
+__all__ = ["ExperienceUpdateParams", "Logo"]
 
 
 class ExperienceUpdateParams(TypedDict, total=False):
@@ -28,26 +28,8 @@ class ExperienceUpdateParams(TypedDict, total=False):
     """The ID of the section to update."""
 
 
-class LogoAttachmentInputWithDirectUploadID(TypedDict, total=False):
-    """Input for an attachment"""
-
-    direct_upload_id: Required[str]
-    """This ID should be used the first time you upload an attachment.
-
-    It is the ID of the direct upload that was created when uploading the file to S3
-    via the mediaDirectUpload mutation.
-    """
-
-
-class LogoAttachmentInputWithID(TypedDict, total=False):
-    """Input for an attachment"""
+class Logo(TypedDict, total=False):
+    """The logo for the experience"""
 
     id: Required[str]
-    """The ID of an existing attachment object.
-
-    Use this when updating a resource and keeping a subset of the attachments. Don't
-    use this unless you know what you're doing.
-    """
-
-
-Logo: TypeAlias = Union[LogoAttachmentInputWithDirectUploadID, LogoAttachmentInputWithID]
+    """The ID of an existing file object."""