+++ src/whop_sdk/types/forum_post_create_params.py
@@ -2,20 +2,13 @@
 
 from __future__ import annotations
 
-from typing import Union, Iterable, Optional
-from typing_extensions import Required, TypeAlias, TypedDict
+from typing import Iterable, Optional
+from typing_extensions import Required, TypedDict
 
 from .shared.currency import Currency
 from .forum_post_visibility_type import ForumPostVisibilityType
 
-__all__ = [
-    "ForumPostCreateParams",
-    "Attachment",
-    "AttachmentAttachmentInputWithDirectUploadID",
-    "AttachmentAttachmentInputWithID",
-    "Poll",
-    "PollOption",
-]
+__all__ = ["ForumPostCreateParams", "Attachment", "Poll", "PollOption"]
 
 
 class ForumPostCreateParams(TypedDict, total=False):
@@ -67,29 +60,11 @@ class ForumPostCreateParams(TypedDict, total=False):
     """The visibility types for forum posts"""
 
 
-class AttachmentAttachmentInputWithDirectUploadID(TypedDict, total=False):
-    """Input for an attachment"""
-
-    direct_upload_id: Required[str]
-    """This ID should be used the first time you upload an attachment.
-
-    It is the ID of the direct upload that was created when uploading the file to S3
-    via the mediaDirectUpload mutation.
-    """
-
-
-class AttachmentAttachmentInputWithID(TypedDict, total=False):
+class Attachment(TypedDict, total=False):
     """Input for an attachment"""
 
     id: Required[str]
-    """The ID of an existing attachment object.
-
-    Use this when updating a resource and keeping a subset of the attachments. Don't
-    use this unless you know what you're doing.
-    """
-
-
-Attachment: TypeAlias = Union[AttachmentAttachmentInputWithDirectUploadID, AttachmentAttachmentInputWithID]
+    """The ID of an existing file object."""
 
 
 class PollOption(TypedDict, total=False):