+++ src/whop_sdk/types/membership_list_response.py
@@ -2,6 +2,7 @@
 
 from typing import Dict, Optional
 from datetime import datetime
+from typing_extensions import Literal
 
 from .._models import BaseModel
 from .shared.currency import Currency
@@ -57,6 +58,9 @@ class User(BaseModel):
     id: str
     """The internal ID of the user."""
 
+    email: Optional[str] = None
+    """The email of the user"""
+
     name: Optional[str] = None
     """The name of the user from their Whop account."""
 
@@ -77,6 +81,16 @@ class MembershipListResponse(BaseModel):
     cycle.
 
     Only applies for memberships that have a renewal plan.
+    """
+
+    cancel_option: Optional[
+        Literal[
+            "too_expensive", "switching", "missing_features", "technical_issues", "bad_experience", "other", "testing"
+        ]
+    ] = None
+    """
+    The different reasons a user can choose for why they are canceling their
+    membership.
     """
 
     canceled_at: Optional[datetime] = None