+++ src/whop_sdk/types/payout_method_list_response.py
@@ -1,6 +1,7 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 from typing import Optional
+from datetime import datetime
 from typing_extensions import Literal
 
 from .._models import BaseModel
@@ -34,8 +35,17 @@ class PayoutMethodListResponse(BaseModel):
     id: str
     """The ID of the payout token"""
 
+    account_reference: Optional[str] = None
+    """
+    A reference to identify the payout destination, such as the last 4 digits of an
+    account number or an email address.
+    """
+
     company: Optional[Company] = None
     """The company associated with the payout token"""
+
+    created_at: datetime
+    """The date and time the payout token was created"""
 
     currency: str
     """The currency code of the payout destination.
@@ -45,6 +55,9 @@ class PayoutMethodListResponse(BaseModel):
 
     destination: Optional[Destination] = None
     """The payout destination associated with the payout token"""
+
+    institution_name: Optional[str] = None
+    """The name of the bank or financial institution."""
 
     is_default: bool
     """Whether this payout token is the default for the payout account"""