+++ tests/api_resources/test_forum_posts.py
@@ -34,7 +34,7 @@ class TestForumPosts:
     def test_method_create_with_all_params(self, client: Whop) -> None:
         forum_post = client.forum_posts.create(
             experience_id="exp_xxxxxxxxxxxxxx",
-            attachments=[{"direct_upload_id": "direct_upload_id"}],
+            attachments=[{"id": "id"}],
             content="content",
             is_mention=True,
             parent_id="parent_id",
@@ -135,7 +135,7 @@ class TestForumPosts:
     def test_method_update_with_all_params(self, client: Whop) -> None:
         forum_post = client.forum_posts.update(
             id="id",
-            attachments=[{"direct_upload_id": "direct_upload_id"}],
+            attachments=[{"id": "id"}],
             content="content",
             is_pinned=True,
             title="title",
@@ -244,7 +244,7 @@ class TestAsyncForumPosts:
     async def test_method_create_with_all_params(self, async_client: AsyncWhop) -> None:
         forum_post = await async_client.forum_posts.create(
             experience_id="exp_xxxxxxxxxxxxxx",
-            attachments=[{"direct_upload_id": "direct_upload_id"}],
+            attachments=[{"id": "id"}],
             content="content",
             is_mention=True,
             parent_id="parent_id",
@@ -345,7 +345,7 @@ class TestAsyncForumPosts:
     async def test_method_update_with_all_params(self, async_client: AsyncWhop) -> None:
         forum_post = await async_client.forum_posts.update(
             id="id",
-            attachments=[{"direct_upload_id": "direct_upload_id"}],
+            attachments=[{"id": "id"}],
             content="content",
             is_pinned=True,
             title="title",