+++ tests/api_resources/test_withdrawals.py
@@ -41,6 +41,8 @@ class TestWithdrawals:
             company_id="biz_xxxxxxxxxxxxxx",
             currency="usd",
             payout_method_id="payout_method_id",
+            platform_covers_fees=True,
+            statement_descriptor="statement_descriptor",
         )
         assert_matches_type(WithdrawalCreateResponse, withdrawal, path=["response"])
 
@@ -189,6 +191,8 @@ class TestAsyncWithdrawals:
             company_id="biz_xxxxxxxxxxxxxx",
             currency="usd",
             payout_method_id="payout_method_id",
+            platform_covers_fees=True,
+            statement_descriptor="statement_descriptor",
         )
         assert_matches_type(WithdrawalCreateResponse, withdrawal, path=["response"])